+++ src/error.ts
@@ -1,3 +1,4 @@
+// tslint:disable:max-classes-per-file
 // tslint:disable-next-line:no-var-requires
 const debug = require("debug").debug("ClientError");
 
@@ -11,3 +12,39 @@ export default class ClientError extends Error {
         this.context = context;
     }
 }
+
+// tslint:disable-next-line:max-classes-per-file
+export class BaseError extends Error {
+    private context?: any;
+
+    constructor(m: string, context?: any) {
+        super(m);
+        this.context = context;
+    }
+}
+
+
+/**
+ * the query limit parameter must be a number larger than 0
+ */
+export class QueryLimitError extends BaseError { };
+
+/**
+ * the query offset parameter must be a number larger than 0
+ */
+export class QueryOffsetError extends BaseError { };
+
+/**
+ * user group already exists
+ */
+export class UserGroupAlreadyExistsError extends BaseError { };
+
+/**
+ * user group does not exist
+ */
+export class UserGroupDoesNotExistError extends BaseError { };
+
+/**
+ * user group cloud not be deleted
+ */
+export class UserGroupDeletionFailedError extends BaseError { };