+++ CHANGELOG.md
@@ -5,6 +5,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- new user group handing on Client and User Group object
+
+### Removed
+- user group handling on client (getGroupsDetails, getGroups, getGroupsDetailsByID)
 
 ## 1.2.1 - 2020-04-24
 - fixed getNotifications